Improve the human readable state of services depending on their type
